added the cardgrid file to keep the gamescreen file small and cleaner so we won't have one file with like 1000 lines of code
